+++ src/main/java/com/bookripple/api/domain/question/service/QuestionServiceImpl.java
@@ -1,17 +1,11 @@
 package com.bookripple.api.domain.question.service;
 
-import com.bookripple.api.common.code.BookErrorCode;
-import com.bookripple.api.common.code.CommonErrorCode;
-import com.bookripple.api.common.code.QuestionErrorCode;
-import com.bookripple.api.common.code.ReadingErrorCode;
-import com.bookripple.api.common.error.ApiException;
-import com.bookripple.api.domain.ai.dto.AiResDto.AiQuestion;
-import com.bookripple.api.domain.ai.enums.AiQuestionType;
-import com.bookripple.api.domain.ai.service.AiService;
+import com.bookripple.api.domain.book.code.BookErrorCode;
 import com.bookripple.api.domain.book.entity.Book;
 import com.bookripple.api.domain.book.repository.BookRepository;
 import com.bookripple.api.domain.member.entity.Member;
 import com.bookripple.api.domain.member.repository.MemberRepository;
+import com.bookripple.api.domain.question.code.QuestionErrorCode;
 import com.bookripple.api.domain.question.converter.QuestionConverter;
 import com.bookripple.api.domain.question.dto.QuestionResDto.MyQ;
 import com.bookripple.api.domain.question.dto.QuestionResDto.MyQuestionList;
@@ -26,12 +20,18 @@ import com.bookripple.api.domain.question.enums.QuestionType;
 import com.bookripple.api.domain.question.repository.QuestionRepository;
 import com.bookripple.api.domain.question.repository.ReadingQuestionRepository;
 import com.bookripple.api.domain.question.repository.SearchQuestionLogRepository;
+import com.bookripple.api.domain.reading.code.ReadingErrorCode;
 import com.bookripple.api.domain.reading.entity.ReadingProgress;
 import com.bookripple.api.domain.reading.repository.ReadingProgressRepository;
+import com.bookripple.api.global.code.CommonErrorCode;
 import com.bookripple.api.global.converter.GlobalConverter;
 import com.bookripple.api.global.dto.GlobalDto.ContentReq;
 import com.bookripple.api.global.dto.GlobalDto.IdList;
 import com.bookripple.api.global.dto.GlobalDto.IdRes;
+import com.bookripple.api.global.error.ApiException;
+import com.bookripple.api.infrastructure.ai.dto.AiResDto.AiQuestion;
+import com.bookripple.api.infrastructure.ai.enums.AiQuestionType;
+import com.bookripple.api.infrastructure.ai.service.AiService;
 import jakarta.transaction.Transactional;
 import java.math.BigDecimal;
 import java.util.List;
